LRCI-3931 Move the initial build entity logic to the base job entity
